+++ src/program/GameLoop.cpp
@@ -60,6 +60,7 @@
 #include <sys/wait.h> // waitpid
 #include <stdint.h>
 #include <cstdlib>
+#include <filesystem>
 
 GameLoop::GameLoop(Context* c) : movie(MovieFile(c)), context(c)
 {
@@ -197,7 +198,7 @@ void GameLoop::init()
         encoding_segment = 0;
 
     /* Extract the game executable name from the game executable path */
-    context->gamename = fileFromPath(context->gamepath);
+    context->gamename = context->gamepath.filename();
 
     /* Clear the event queue and parameters */
     gameEvents->init();
@@ -379,24 +380,28 @@ void GameLoop::initProcessMessages()
         int index = 0;
         sendData(&index, sizeof(int));
 
-        std::string basesavestatepath = context->config.savestatedir + '/';
-        basesavestatepath += context->gamename;
+        std::filesystem::path basesavestatepath = context->config.savestatedir;
+        basesavestatepath /= context->gamename;
         basesavestatepath += ".state0";
         sendMessage(MSGN_BASE_SAVESTATE_PATH);
-        sendString(basesavestatepath);
+        sendString(basesavestatepath.string());
     }
 
     /* Send the Steam user data path and remote storage */
     if (context->config.sc.virtual_steam) {
         sendMessage(MSGN_STEAM_USER_DATA_PATH);
         sendString(context->config.steamuserdir);
-        std::string remotestorage = context->config.steamuserdir + "/";
-        remotestorage += context->gamename;
-        if (create_dir(remotestorage) < 0) {
+        std::filesystem::path remotestorage = context->config.steamuserdir;
+        remotestorage /= context->gamename;
+        try {
+            std::filesystem::create_directory(remotestorage);
+        }
+        catch (std::filesystem::filesystem_error const& ex) {
             std::cerr << "Cannot create dir " << remotestorage << std::endl;
+            std::cerr << "what():  " << ex.what() << std::endl;
         }
         sendMessage(MSGN_STEAM_REMOTE_STORAGE);
-        sendString(remotestorage);
+        sendString(remotestorage.string());
     }
 
     sendMessage(MSGN_ENCODING_SEGMENT);