+++ src/program/utils.h
@@ -22,21 +22,10 @@
 
 #include <string>
 #include <cstdint>
+#include <filesystem>
 
 /* Forward declaration */
 struct Context;
-
-/* Get the file from path */
-std::string fileFromPath(const std::string& path);
-
-/* Get the directory from path */
-std::string dirFromPath(const std::string& path);
-
-/* Get the absolute path of a file even if it doesn't exist */
-std::string realpath_nonexist(const std::string& path);
-
-/* Create a directory if it does not exist already */
-int create_dir(const std::string& path);
 
 /* Remove savestate files */
 void remove_savestates(Context* context);
@@ -59,10 +48,10 @@ enum BinaryType {
 };
 
 /* Run the `file` command from a shell and extract the output of the command. */
-int extractBinaryType(std::string path);
+int extractBinaryType(std::filesystem::path path);
 
 /* Get the executable from MacOS .app directory. */
-std::string extractMacOSExecutable(std::string path);
+std::filesystem::path extractMacOSExecutable(std::filesystem::path path);
 
 /* Get the result of a shell command */
 std::string queryCmd(const std::string& cmd, int* status = nullptr);